Refactor and reorganize
